+++ src/main/java/com/rwtema/extrautils/item/filters/Matcher.java
@@ -11,38 +11,38 @@ import net.minecraftforge.oredict.OreDictionary;
 
 public class Matcher {
   public Type type;
-  
+
   public final String name;
-  
+
   public final String unlocalizedName;
-  
+
   int index;
-  
+
   private boolean addToNEI;
-  
+
   public static String getUnlocalizedName(String name) {
     name = name.replaceAll("[^{A-Za-z0-9}]", "").toLowerCase(Locale.ENGLISH);
     return "item.extrautils:nodeUpgrade.10.program." + name;
   }
-  
+
   private static String titleCase(String prefix) {
     return prefix.substring(0, 1).toUpperCase(Locale.ENGLISH) + prefix.substring(1, prefix.length());
   }
-  
+
   public String getLocalizedName() {
     if (!StatCollector.canTranslate(this.unlocalizedName))
-      return this.name + ".exe"; 
+      return this.name + ".exe";
     return StatCollector.translateToLocal(this.unlocalizedName);
   }
-  
+
   public boolean isSelectable() {
     return true;
   }
-  
+
   public Matcher(String name) {
     this(name, true);
   }
-  
+
   public Matcher(String name, boolean addToNEI) {
     this.name = name;
     this.addToNEI = addToNEI;
@@ -63,10 +63,10 @@ public class Matcher {
       case 3:
         type = Type.BOTH;
         break;
-    } 
+    }
     this.type = type;
   }
-  
+
   public boolean methodExists(String method, Class<?> clazz, Class... classes) {
     try {
       clazz.getDeclaredMethod(method, classes);
@@ -74,160 +74,160 @@ public class Matcher {
     } catch (NoSuchMethodException e) {
       clazz = clazz.getSuperclass();
       return (clazz != Matcher.class && methodExists(method, clazz, classes));
-    } 
+    }
   }
-  
+
   public boolean matchFluid(FluidStack fluid) {
     return false;
   }
-  
+
   public boolean matchItem(ItemStack item) {
     return false;
   }
-  
+
   public boolean shouldAddToNEI() {
     return this.addToNEI;
   }
-  
+
   public enum Type {
     FLUID, ITEM, BOTH;
   }
-  
+
   public static class MatcherTool extends Matcher {
     private final String tool;
-    
+
     public MatcherTool(String tool) {
       super("Tool" + Matcher.titleCase(tool));
       this.tool = tool;
     }
-    
+
     public boolean matchItem(ItemStack item) {
       for (String s : item.getItem().getToolClasses(item)) {
         if (this.tool.equals(s))
-          return true; 
-      } 
+          return true;
+      }
       return false;
     }
   }
-  
+
   public static class MatcherOreDic extends Matcher {
     private final String prefix;
-    
+
     private final TIntByteHashMap map = new TIntByteHashMap();
-    
+
     public MatcherOreDic(String prefix) {
       super("OreDic" + Matcher.titleCase(prefix));
       this.prefix = prefix;
     }
-    
+
     public boolean matchItem(ItemStack item) {
       for (int i : OreDictionary.getOreIDs(item)) {
         if (this.map.containsKey(i)) {
           if (this.map.get(i) != 0)
-            return true; 
+            return true;
         } else {
           this.map.put(i, (byte)(OreDictionary.getOreName(i).startsWith(this.prefix) ? 1 : 0));
-        } 
-      } 
+        }
+      }
       return false;
     }
-    
+
     public boolean isSelectable() {
       for (String s : OreDictionary.getOreNames()) {
         if (s.startsWith(this.prefix))
-          return true; 
-      } 
+          return true;
+      }
       return false;
     }
   }
-  
+
   public static abstract class MatcherItem extends Matcher {
     HashSet<Item> entries;
-    
+
     public MatcherItem(String name) {
       super(name);
     }
-    
+
     public void buildMap() {
       this.entries = new HashSet<Item>();
       for (Object anItemRegistry : Item.itemRegistry) {
         Item item = (Item)anItemRegistry;
         if (matchItem(item))
-          this.entries.add(item); 
-      } 
+          this.entries.add(item);
+      }
     }
-    
+
     public boolean matchItem(ItemStack item) {
       if (this.entries == null)
-        buildMap(); 
+        buildMap();
       return this.entries.contains(item.getItem());
     }
-    
+
     public boolean isSelectable() {
       if (this.entries == null)
-        buildMap(); 
+        buildMap();
       return !this.entries.isEmpty();
     }
-    
+
     protected abstract boolean matchItem(Item param1Item);
   }
-  
+
   public static class MatcherOreDicPair extends Matcher {
     private final String prefix;
-    
+
     private final String prefix2;
-    
+
     private final TIntByteHashMap map = new TIntByteHashMap();
-    
+
     public MatcherOreDicPair(String prefix, String prefix2) {
       super("OrePair" + Matcher.titleCase(prefix) + Matcher.titleCase(prefix2));
       this.prefix = prefix;
       this.prefix2 = prefix2;
     }
-    
+
     public boolean matchItem(ItemStack item) {
       for (int i : OreDictionary.getOreIDs(item)) {
         String oreName = OreDictionary.getOreName(i);
         boolean isOre = oreName.startsWith(this.prefix);
         isOre = (isOre && oreExists(oreName.replaceFirst(this.prefix, this.prefix2)));
-        this.map.put(i, isOre ? 1 : 0);
+        this.map.put(i, isOre ? (byte)1 : (byte)0);
         if (isOre)
-          return true; 
-      } 
+          return true;
+      }
       return false;
     }
-    
+
     public boolean oreExists(String k) {
       for (String s : OreDictionary.getOreNames()) {
-        if (k.equals(s) && 
+        if (k.equals(s) &&
           !OreDictionary.getOres(k, false).isEmpty())
-          return true; 
-      } 
+          return true;
+      }
       return false;
     }
-    
+
     public boolean isSelectable() {
       for (String s : OreDictionary.getOreNames()) {
         if (s.startsWith(this.prefix) && oreExists(s.replace(this.prefix, this.prefix2)))
-          return true; 
-      } 
+          return true;
+      }
       return false;
     }
   }
-  
+
   public static class InverseMatch extends Matcher {
     private final Matcher matcher;
-    
+
     public InverseMatch(String item, Matcher matcher) {
       super(item);
       this.matcher = matcher;
       this.type = matcher.type;
     }
-    
+
     public boolean matchItem(ItemStack item) {
       return !this.matcher.matchItem(item);
     }
-    
+
     public boolean matchFluid(FluidStack fluid) {
       return !this.matcher.matchFluid(fluid);
     }
